move clientside network files to network folder
