+++ server/server.lua
@@ -6,7 +6,7 @@ require("class")
 json = require("libraries.dkjson")
 require("server.stridx")
 require("csprng")
-local NetworkProtocol = require("NetworkProtocol")
+local NetworkProtocol = require("network.NetworkProtocol")
 require("server.server_globals")
 require("server.server_file_io")
 require("server.Connection")
